Sweet!

+++ include/estimator.h
@@ -5,10 +5,9 @@
 
 typedef struct
 {
-  int16_t p;
-  int16_t q;
-  int16_t r;
-
+  int32_t p;
+  int32_t q;
+  int32_t r;
   int32_t phi;
   int32_t theta;
   int32_t psi;
